pyproject.toml
[tool.poetry]
name = "analiz bankovskikh operatsii"
version = "0.1.0"
description = "Приложение для анализа транзакций"
authors = ["Aleksandrina Anna kurai@vk.com"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
openpyxl = "^3.1.0"
requests = "^2.31.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.0"
types-requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119
exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
)/
'''

[tool.isort]
profile = "black"
line_length = 119
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = ["venv", "\\.venv"]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"
